+++ CHANGES.rst
@@ -14,6 +14,12 @@ Changelog
 
 -   Update bilby_pipe to 1.1.2.
 
+-   Updated RapidPE/RIFT Pipeline to 0.5.3, replacing errors from absent GstLAL
+    triggers with a graceful failure.  Updated RIFT to 0.0.15.9 to fix
+    compatibility with newer NumPy versions.  Updated LALSuite to 7.16 to fix
+    bug in `XLALSimInspiralTDModesFromPolarizations` causing RIFT to fail.
+
+
 2.1.3 "Thunderbird" (2023-06-24)
 --------------------------------